requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
certifi==2023.11.17
charset-normalizer==3.3.2
construct==2.10.68
construct-typing==0.5.6
cytoolz==0.12.2
Django==5.0.1
django-filter==24.2
djangorestframework==3.15.1
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.0
eth-typing==3.5.2
eth-utils==2.3.1
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.2
httpx==0.25.2
idna==3.6
jsonalias==0.1.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
Markdown==3.6
multidict==6.0.4
numpy==1.26.4
oauthlib==3.2.2
parsimonious==0.9.0
protobuf==4.25.1
psutil==5.9.8
psycopg2==2.9.9
pycryptodome==3.19.1
pyTelegramBotAPI==4.14.1
python-decouple==3.8
python-telegram-bot==20.7
pyunormalize==15.1.0
pywin32==306
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rlp==4.0.0
rpds-py==0.16.2
sniffio==1.3.0
solana==0.33.0
solders==0.21.0
sortedcontainers==2.4.0
sqlparse==0.4.4
toolz==0.12.0
tweepy==4.14.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
validators==0.22.0
watchdog==4.0.0
web3==6.13.0
websockets==11.0.3
yarl==1.9.4
